code clean up and modernization
